% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenization.R
\name{apply_to_chars}
\alias{apply_to_chars}
\title{Apply a function to each character in a string.}
\usage{
apply_to_chars(text, .f, ...)
}
\arguments{
\item{text}{A character scalar to process.}

\item{.f}{The function to apply to each character. Should return a character
scalar, given a single-character input.}

\item{...}{Other arguments to pass to .f.}
}
\value{
The character scalar obtained by applying the given function to
each character of the input string, and concatenating the results.
}
\description{
Utility function for something done a lot in this package.
}
\examples{
apply_to_chars("some words",
               function(char) {
                 if (char \%in\% c("a", "e", "o")) {
                   return(paste0(char, char))
                 }
                 return(char)
               })
}
\keyword{internal}
