% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{assert_rank}
\alias{assert_rank}
\title{Confirm the rank of a tensor}
\usage{
assert_rank(tensor, expected_rank, name = NULL)
}
\arguments{
\item{tensor}{A tf.Tensor to check the rank of.}

\item{expected_rank}{Integer vector or list of integers, expected rank.}

\item{name}{Optional name of the tensor for the error message.}
}
\value{
TRUE if the Tensor is of the expected rank (error otherwise).
}
\description{
Throws an error if the tensor rank is not of the expected rank.
}
\examples{
\dontrun{
with(tensorflow::tf$variable_scope("examples",
  reuse = tensorflow::tf$AUTO_REUSE
), {
  ids <- tensorflow::tf$get_variable("x", dtype = "int32", shape = 10L)
  assert_rank(ids, 1)
  assert_rank(ids, 1:2)
  assert_rank(ids, 2)
})
}
}
