% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{bert_config_from_json_file}
\alias{bert_config_from_json_file}
\title{Load BERT config object from json file}
\usage{
bert_config_from_json_file(json_file)
}
\arguments{
\item{json_file}{Character; the path to a json config file.}
}
\value{
An object of class BertConfig
}
\description{
Given a path to a json config file, construct a BertConfig object with
appropriate values.
}
\examples{
\dontrun{
temp_dir <- tempdir()
json_file <- file.path(
  temp_dir,
  "BERT_checkpoints",
  "uncased_L-12_H-768_A-12",
  "bert_config.json"
)
bert_config_from_json_file(json_file)
}
}
