% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenization.R
\name{check_vocab}
\alias{check_vocab}
\title{Check Vocabulary}
\usage{
check_vocab(words, ckpt_dir = NULL, vocab_file = find_vocab(ckpt_dir))
}
\arguments{
\item{words}{Character vector; words to check.}
}
\value{
A logical vector containing \code{TRUE} if the corresponding word was
  found verbatim in the vocabulary, \code{FALSE} otherwise.
}
\description{
Given some words and a word piece vocabulary, checks to see if the words are
in the vocabulary.
}
\examples{
\dontrun{
BERT_PRETRAINED_DIR <- download_BERT_checkpoint("bert_base_uncased")
to_check <- c("apple", "appl")
check_vocab(words = to_check, ckpt_dir = BERT_PRETRAINED_DIR) # TRUE, FALSE
#'
}
}
