% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenization.R
\name{clean_text}
\alias{clean_text}
\title{Perform invalid character removal and whitespace cleanup on text.}
\usage{
clean_text(text)
}
\arguments{
\item{text}{A character scalar.}
}
\value{
Cleaned up text.
}
\description{
(R implementation of BasicTokenizer._clean_text from
BERT: tokenization.py.)
}
\keyword{internal}
