% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenization.R
\name{convert_by_vocab}
\alias{convert_by_vocab}
\alias{convert_tokens_to_ids}
\alias{convert_ids_to_tokens}
\title{Convert a sequence of tokens/ids using the provided vocab.}
\usage{
convert_by_vocab(vocab, items)

convert_tokens_to_ids(vocab, tokens)

convert_ids_to_tokens(inv_vocab, ids)
}
\arguments{
\item{vocab}{Vocabulary; provides mapping from index to tokens. (This may
be in fact an "inverse vocabulary", where the names are the indices and
the values are the tokens.)}

\item{items}{Vector of the keys (names in the vocab vector) to "convert".}

\item{tokens}{Equivalent to items.}

\item{inv_vocab}{Equivalent to vocab.}

\item{ids}{Equivalent to items.}
}
\value{
Vector of the values in `vocab` corresponding to `items`.
(The names on the returned vector are kept.)
}
\description{
Convert a sequence of tokens/ids using the provided vocab.
}
\section{Functions}{
\itemize{
\item \code{convert_tokens_to_ids}: Wrapper function for specifically converting
tokens to ids.

\item \code{convert_ids_to_tokens}: Wrapper function for specifically converting
ids to tokens.
}}

\examples{
convert_by_vocab(c("token1"=0, "token2"=1), "token1")
}
