% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_classifier.R
\name{convert_examples_to_features}
\alias{convert_examples_to_features}
\title{Convert \code{InputExample}s to \code{InputFeatures}}
\usage{
convert_examples_to_features(examples, label_list, max_seq_length, tokenizer)
}
\arguments{
\item{examples}{List of \code{InputExample}s to convert.}

\item{label_list}{Character (or integer?); possible labels for examples.}

\item{max_seq_length}{Integer; the maximum number of tokens that will be
considered together.}

\item{tokenizer}{A tokenizer object to use (e.g. object of class
FullTokenizer).}
}
\value{
A list of \code{InputFeatures}.
}
\description{
Converts a set of \code{InputExample}s to a list of \code{InputFeatures}.
}
\examples{
\dontrun{
tokenizer <- FullTokenizer("vocab.txt")
input_ex1 <- InputExample(guid = 1L,
                         text_a = "Some text to classify.",
                         text_b = "More wordy words.",
                         label = "good")
input_ex2 <- InputExample(guid = 2L,
                         text_a = "This is another example.",
                         text_b = "So many words.",
                         label = "bad")
feat <- convert_examples_to_features(examples = list(input_ex1, input_ex2),
                                     label_list = c("good", "bad"),
                                     max_seq_length = 15L,
                                     tokenizer = tokenizer)
}
}
