% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_classifier.R
\name{convert_single_example}
\alias{convert_single_example}
\title{Convert a single \code{InputExample} into a single \code{InputFeatures}}
\usage{
convert_single_example(ex_index, example, label_list, max_seq_length,
  tokenizer)
}
\arguments{
\item{ex_index}{Integer; the index of this example. This is used to determine
whether or not to print out some log info (for debugging or runtime
confirmation). It is assumed this starts with 1 (in R).}

\item{example}{The \code{InputExample} to convert.}

\item{label_list}{Character (or integer); allowed labels for these examples.}

\item{max_seq_length}{Integer; the maximum number of tokens that will be
considered together.}

\item{tokenizer}{A tokenizer object to use (e.g. object of class
FullTokenizer).}
}
\value{
An object of class \code{InputFeatures}.
}
\description{
Converts a single \code{InputExample} into a single \code{InputFeatures}.
}
\examples{
\dontrun{
tokenizer <- FullTokenizer("vocab.txt")
input_ex <- InputExample(guid = 1L,
                         text_a = "Some text to classify.",
                         text_b = "More wordy words.",
                         label = "good")
feat <- convert_single_example(ex_index = 1L,
                               example = input_ex,
                               label_list = c("good", "bad"),
                               max_seq_length = 15L,
                               tokenizer = tokenizer)
}
}
