% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_features.R
\name{convert_single_example_EF}
\alias{convert_single_example_EF}
\title{Convert a single \code{InputExample_EF} into a single \code{InputFeatures_EF}}
\usage{
convert_single_example_EF(ex_index, example, seq_length, tokenizer)
}
\arguments{
\item{ex_index}{Integer; the index of this example. This is used to determine
whether or not to print out some log info (for debugging or runtime
confirmation). It is assumed this starts with 1 (in R).}

\item{example}{The \code{InputExample_EF} to convert.}

\item{seq_length}{Integer; the maximum number of tokens that will be
considered together.}

\item{tokenizer}{A tokenizer object to use (e.g. object of class
FullTokenizer).}
}
\value{
An object of class \code{InputFeatures_EF}.
}
\description{
Converts a single \code{InputExample_EF} into a single
\code{InputFeatures_EF}. Very similar to \code{convert_single_example} from
run_classifier.R. (RBERT issue #28.)
}
\keyword{internal}
