% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenization.R
\name{convert_to_unicode}
\alias{convert_to_unicode}
\title{Convert `text` to Unicode}
\usage{
convert_to_unicode(text)
}
\arguments{
\item{text}{character scalar to convert to unicode}
}
\value{
input text, converted to unicode if applicable
}
\description{
See documentation for `Encoding` for more information.
Assumes utf-8 input.
}
\examples{
convert_to_unicode("fa\xC3\xA7ile")
}
