% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{create_initializer}
\alias{create_initializer}
\title{Create truncated normal initializer}
\usage{
create_initializer(initializer_range = 0.02)
}
\arguments{
\item{initializer_range}{A double describing the range for the initializer
(passed to the stddev parameter).}
}
\value{
A tensorflow initializer.
}
\description{
This is a wrapper around the tensorflow truncated_normal_initializer
function.
}
\examples{
\dontrun{
create_initializer(0.02)
}
}
