% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf2-layer_bert.R
\name{custom_layer_BERT}
\alias{custom_layer_BERT}
\title{Custom Layer: BERT model}
\usage{
custom_layer_BERT(
  object,
  name = NULL,
  trainable = NULL,
  param_list = list(),
  ...
)
}
\arguments{
\item{object}{Model or layer object.}

\item{name}{Character; An optional name for the layer. Must be unique in a
model.}

\item{trainable}{Logical; whether the layer weights will be updated during
training.}

\item{param_list}{A named list of parameter values used in defining the
layer. \emph{standardize this here}
\describe{
\item{\code{xxx}}{description }
}}
}
\description{
Create a BERT (or ALBERT) model.
}
