% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf2-layer_embeddings.R
\name{custom_layer_position_embedding}
\alias{custom_layer_position_embedding}
\title{Custom Layer: Position Embedding}
\usage{
custom_layer_position_embedding(
  object,
  name = NULL,
  trainable = NULL,
  param_list = list(),
  ...
)
}
\arguments{
\item{object}{Model or layer object.}

\item{name}{Character; An optional name for the layer. Must be unique in a
model.}

\item{trainable}{Logical; whether the layer weights will be updated during
training.}

\item{param_list}{A named list of parameter values used in defining the
layer.
\describe{
\item{\code{hidden_size}}{Integer; The size of the output. Should match the
size of the second input.}
\item{\code{max_position_embeddings}}{Numeric; .}
\item{\code{initializer_range}}{Numeric; the value passed in as the \code{stddev}
parameter to the \code{initializer_truncated_normal} in the dense layer
initializer.}
\item{\code{dtype}}{The data type of the layer output. Defaults to "float32".
Valid values from \code{tensorflow::tf$float32$name}, etc. }
}}
}
\description{
Create position embeddings for the input layer.
}
