% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf2-layer_transformer.R
\name{custom_layer_transformer_encoder}
\alias{custom_layer_transformer_encoder}
\title{Custom Layer: Transformer Encoder Layer}
\usage{
custom_layer_transformer_encoder(
  object,
  name = NULL,
  trainable = NULL,
  param_list = list(),
  ...
)
}
\arguments{
\item{object}{Model or layer object.}

\item{name}{Character; An optional name for the layer. Must be unique in a
model.}

\item{trainable}{Logical; whether the layer weights will be updated during
training.}

\item{param_list}{A named list of parameter values used in defining the
layer.
\describe{
\item{\code{dtype}}{The data type of the layer output.
Defaults to "float32". Valid values from \code{tensorflow::tf$float32$name},
etc. }
}}
}
\description{
Create a multi-layer transformer-based encoder.
}
