% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_checkpoint.R
\name{.choose_BERT_dir}
\alias{.choose_BERT_dir}
\title{Choose a directory for BERT checkpoints}
\usage{
.choose_BERT_dir(dir)
}
\arguments{
\item{dir}{Character vector. Destination directory for checkpoint. Leave
`NULL` to allow `RBERT` to automatically choose a directory. The path is
determined from the `dir` parameter if supplied, followed by the
`RBERT.dir` option (set using \link{set_BERT_dir}), followed by an "RBERT"
folder in the user cache directory (determined using
\code{\link[rappdirs]{user_cache_dir()}}). If you provide a `dir`, the
`RBERT.dir` option will be updated to that location.}
}
\value{
A character vector indicating a directory in which BERT checkpoints
  are stored.
}
\description{
If `dir` is not NULL, this function simply returns `dir`. Otherwise it checks
the `RBERT.dir` param, and then uses \code{\link[rappdirs]{user_cache_dir}}
to choose a directory if necessary.
}
\keyword{internal}
