% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_features.R
\name{.convert_examples_to_features_EF}
\alias{.convert_examples_to_features_EF}
\title{Convert \code{InputExample_EF}s to \code{InputFeatures_EF}}
\usage{
.convert_examples_to_features_EF(examples, seq_length, tokenizer)
}
\arguments{
\item{examples}{List of \code{InputExample_EF}s to convert.}

\item{seq_length}{Integer; the maximum number of tokens that will be
considered together.}

\item{tokenizer}{A tokenizer object to use (e.g. object of class
FullTokenizer).}
}
\value{
A list of \code{InputFeatures}.
}
\description{
Converts a set of \code{InputExample_EF}s to a list of
\code{InputFeatures_EF}. Very similar to \code{convert_examples_to_features}
from run_classifier.R. (RBERT issue #28.)
}
\keyword{internal}
