% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf2-utils.R
\name{.create_attention_mask}
\alias{.create_attention_mask}
\title{Create an Attention Mask}
\usage{
.create_attention_mask(from_shape, input_mask)
}
\arguments{
\item{from_shape}{List or integer vector. First two elements should be
\code{batch_size} and \code{from_seq_len}.}

\item{input_mask}{Tensor with dimensions \code{[batch_size, seq_len]}.}
}
\value{
A tensor of ones with dimensions \code{[batch_size, from_seq_len,
  seq_len]}.
}
\description{
Creates a 3D attention mask.
}
\details{
Attention masks are used to control which words can pay attention to which.
This function just sets up the structure of the mask (a matrix of dimensions
From x To for each incoming example in the batch), and populates it with
ones.
}
\keyword{internal}
