% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_checkpoint.R
\name{.download_BERT_checkpoint}
\alias{.download_BERT_checkpoint}
\title{Download a checkpoint zip file}
\usage{
.download_BERT_checkpoint(url, checkpoint_zip_path)
}
\arguments{
\item{url}{Character vector. An optional url from which to download a
checkpoint. Overrides \code{model} parameter if not NULL.}

\item{checkpoint_zip_path}{The path to which the checkpoint zip should be
downloaded.}
}
\value{
\code{TRUE} invisibly.
}
\description{
Download a checkpoint zip file
}
\keyword{internal}
