% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_features.R
\name{.extract_attention_df}
\alias{.extract_attention_df}
\title{Tidy Attention Probabilities}
\usage{
.extract_attention_df(attention_probs)
}
\arguments{
\item{attention_probs}{Raw attention probabilities.}
}
\value{
A tibble of attention weights.
}
\description{
Tidy Attention Probabilities
}
\keyword{internal}
