% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_features.R
\name{.extract_attention_labels}
\alias{.extract_attention_labels}
\title{Tidy Token Labels, Etc}
\usage{
.extract_attention_labels(attention_probs)
}
\arguments{
\item{attention_probs}{Raw attention probabilities.}
}
\value{
A tibble with token_index, token, sequence_index, and segment_index.
}
\description{
Tidy Token Labels, Etc
}
\keyword{internal}
