% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_features.R
\name{.extract_attention_layer_names}
\alias{.extract_attention_layer_names}
\title{Tidy Attention Layer Names}
\usage{
.extract_attention_layer_names(attention_probs)
}
\arguments{
\item{attention_probs}{Raw attention probabilities.}
}
\value{
A tibble of attention layer indexes and fake indexes (a temporary
  index based on this layer's position in the list).
}
\description{
Tidy Attention Layer Names
}
\keyword{internal}
