% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_features.R
\name{.extract_attention_weights}
\alias{.extract_attention_weights}
\title{Tidy Attention Weights}
\usage{
.extract_attention_weights(attention_probs)
}
\arguments{
\item{attention_probs}{Raw attention probabilities.}
}
\value{
A tibble of attention weights
}
\description{
Tidy Attention Weights
}
\keyword{internal}
