% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_features.R
\name{.extract_output_df}
\alias{.extract_output_df}
\title{Extract Embedding Vectors}
\usage{
.extract_output_df(layer_outputs)
}
\arguments{
\item{layer_outputs}{The \code{layer_outputs} component.}
}
\value{
The embedding vector components as a tbl_df, for all tokens and all
  layers.
}
\description{
Extract the embedding vector values for output of
\code{\link{extract_features}}. The resulting tbl_df will typically have a
large number of columns (> 768), so it will be rather slow to
\code{\link{View}}. Consider using \code{\link[dplyr]{glimpse}} if you just
want to peek at the values.
}
\keyword{internal}
