% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_features.R
\name{.extract_output_labels}
\alias{.extract_output_labels}
\title{Extract Labels for Embeddings}
\usage{
.extract_output_labels(layer_outputs)
}
\arguments{
\item{layer_outputs}{The \code{layer_outputs} component.}
}
\value{
The embedding vector components as a tbl_df, for all tokens and all
  layers.
}
\description{
Extract the label columns for embedding vector values for output of
\code{\link{extract_features}}.
}
\keyword{internal}
