% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_features.R
\name{.extract_output_values}
\alias{.extract_output_values}
\title{Extract Embeddings}
\usage{
.extract_output_values(layer_outputs)
}
\arguments{
\item{layer_outputs}{The \code{layer_outputs} component.}
}
\value{
The embedding vector components as a tbl_df, for all tokens and all
  layers.
}
\description{
Extract the embedding vector values from output for
\code{\link{extract_features}}. The columns identifying example sequence,
segment, token, and row are extracted separately, by
\code{\link{.extract_labels}}.
}
\keyword{internal}
