% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_features.R
\name{.get_actual_index}
\alias{.get_actual_index}
\title{Standardize Indices}
\usage{
.get_actual_index(index, length)
}
\arguments{
\item{index}{Integer; the index to normalize.}

\item{length}{Integer; the length of the vector or list we are indexing.}
}
\value{
The "actual" integer index, between 1 and \code{length}, inclusive.
}
\description{
Convert negative indices to positive ones. Use the convention that
\code{vec[[-1L]]} signifies the last element of \code{vec}, \code{vec[[-2L]]}
signifies the second-to-last element of \code{vec}, and so on. 1-based
indexing is assumed. Values of zero, or out-of-range indices, will be
rejected.
}
\keyword{internal}
