% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_checkpoint.R
\name{.get_model_archive_path}
\alias{.get_model_archive_path}
\title{Locate an archive file for a BERT checkpoint}
\usage{
.get_model_archive_path(model, dir, archive_type)
}
\arguments{
\item{model}{Character vector. Which model checkpoint to download.}

\item{dir}{Character vector. Destination directory for checkpoints. Leave
\code{NULL} to allow RBERT to automatically choose a directory. The path is
determined from the \code{dir} parameter if supplied, followed by the
`RBERT.dir` option (set using \link{set_BERT_dir}), followed by an "RBERT"
folder in the user cache directory (determined using
\code{\link[rappdirs]{user_cache_dir}}). If you provide a \code{dir}, the
`RBERT.dir` option will be updated to that location. Note that the
checkpoint will create a subdirectory inside this \code{dir}.}

\item{archive_type}{How is the checkpoint archived? We currently support
"zip" and "tar-gzip". Leave NULL to infer from the \code{url}.}
}
\value{
The path to the archive file where the raw checkpoint should be
  saved.
}
\description{
Locate an archive file for a BERT checkpoint
}
\keyword{internal}
