% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_checkpoint.R
\name{.get_model_archive_type}
\alias{.get_model_archive_type}
\title{Get archive type of a BERT checkpoint}
\usage{
.get_model_archive_type(model)
}
\arguments{
\item{model}{Character vector. Which model checkpoint to download.}
}
\value{
The archive type to the specified BERT model.
}
\description{
Returns the archive type ("zip" or "tar-gzip") of the specified BERT
checkpoint from the Google Research collection or other repository.
}
\keyword{internal}
