% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_checkpoint.R
\name{.get_model_subdir}
\alias{.get_model_subdir}
\title{Locate a subdir for a BERT checkpoint}
\usage{
.get_model_subdir(model, dir)
}
\arguments{
\item{model}{Character vector. Which model checkpoint to download.}

\item{dir}{Character vector. Destination directory for checkpoints. Leave
\code{NULL} to allow RBERT to automatically choose a directory. The path is
determined from the \code{dir} parameter if supplied, followed by the
`RBERT.dir` option (set using \link{set_BERT_dir}), followed by an "RBERT"
folder in the user cache directory (determined using
\code{\link[rappdirs]{user_cache_dir}}). If you provide a \code{dir}, the
`RBERT.dir` option will be updated to that location. Note that the
checkpoint will create a subdirectory inside this \code{dir}.}
}
\value{
The path to the sub-directory where the checkpoint should be saved.
}
\description{
Locate a subdir for a BERT checkpoint
}
\keyword{internal}
