% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_checkpoint.R
\name{.get_model_url}
\alias{.get_model_url}
\title{Get url of a BERT checkpoint}
\usage{
.get_model_url(model)
}
\arguments{
\item{model}{Character vector. Which model checkpoint to download.}
}
\value{
The url to the specified BERT model.
}
\description{
Returns the url of the specified BERT checkpoint from the Google Research
collection or other repository.
}
\keyword{internal}
