% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenization.R
\name{.get_token_type_ids}
\alias{.get_token_type_ids}
\title{Get Token Type IDs from Tokenization}
\usage{
.get_token_type_ids(tokenization, sep_index)
}
\arguments{
\item{tokenization}{Output from \code{tokenize_input}.}

\item{sep_index}{Named integer value, as used within \code{tokenize_input}.}
}
\value{
Integer vector of token type ids.
}
\description{
The token type is 0 for the first segement, and 1 for the second segment.
Segments are determined here by the SEP tokens.
}
\keyword{internal}
