% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_checkpoint.R
\name{.has_checkpoint}
\alias{.has_checkpoint}
\title{Check whether the user already has a checkpoint}
\usage{
.has_checkpoint(model = c("bert_base_uncased", "bert_base_cased",
  "bert_large_uncased", "bert_large_cased", "bert_large_uncased_wwm",
  "bert_large_cased_wwm", "bert_base_multilingual_cased",
  "bert_base_chinese"), dir = NULL, url = NULL)
}
\arguments{
\item{model}{Character vector. Which model checkpoint to download.}

\item{dir}{Character vector. Destination directory for checkpoint. Leave
`NULL` to allow `RBERT` to automatically choose a directory. The path is
determined from the `dir` parameter if supplied, followed by the
`RBERT.dir` option (set using \link{set_BERT_dir}), followed by an "RBERT"
folder in the user cache directory (determined using
\code{\link[rappdirs]{user_cache_dir()}}). If you provide a `dir`, the
`RBERT.dir` option will be updated to that location.}

\item{url}{Character vector. An optional url from which to download a
checkpoint. Overrides \code{model} parameter if not NULL.}
}
\value{
A logical indicating whether the user already has that checkpoint in
  that location.
}
\description{
Check the specified dir (or the default dir if none is specified) for a given
model or url.
}
\keyword{internal}
