% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_checkpoint.R
\name{.infer_archive_type}
\alias{.infer_archive_type}
\title{Infer the archive type for a BERT checkpoint}
\usage{
.infer_archive_type(url)
}
\arguments{
\item{url}{Character vector. An optional url from which to download a
checkpoint. Overrides \code{model} parameter if not NULL.}
}
\value{
A character vector, currently either "zip" or "tar-gzip".
}
\description{
Infer the archive type for a BERT checkpoint
}
\keyword{internal}
