% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_checkpoint.R
\name{.infer_checkpoint_archive_path}
\alias{.infer_checkpoint_archive_path}
\title{Infer the path to the archive for a BERT checkpoint}
\usage{
.infer_checkpoint_archive_path(url, dir)
}
\arguments{
\item{url}{Character vector. An optional url from which to download a
checkpoint. Overrides \code{model} parameter if not NULL.}

\item{dir}{Character vector. Destination directory for checkpoints. Leave
`NULL` to allow `RBERT` to automatically choose a directory. The path is
determined from the `dir` parameter if supplied, followed by the
`RBERT.dir` option (set using \link{set_BERT_dir}), followed by an "RBERT"
folder in the user cache directory (determined using
\code{\link[rappdirs]{user_cache_dir}}). If you provide a `dir`, the
`RBERT.dir` option will be updated to that location. Note that the
checkpoint will create a subdirectory inside this `dir`.}
}
\value{
A character vector file path, pointing to where the raw checkpoint
  archive should be saved.
}
\description{
Infer the path to the archive for a BERT checkpoint
}
\keyword{internal}
