% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_features.R
\name{.InputFeatures_EF}
\alias{.InputFeatures_EF}
\title{Construct objects of class \code{InputFeatures_FE}}
\usage{
.InputFeatures_EF(unique_id, tokens, input_ids, input_mask, input_type_ids)
}
\arguments{
\item{unique_id}{Integer or character; a unique id for this example.}

\item{tokens}{Character vector; the actual tokens in this example.}

\item{input_ids}{Integer vector; the sequence of token ids in this example.}

\item{input_mask}{Integer vector; sequence of 1s (for "real" tokens) and 0s
(for padding tokens).}

\item{input_type_ids}{Integer vector; aka token_type_ids. Indicators for
which sentence (or sequence) each token belongs to. Classical BERT supports
only 0s and 1s (for first and second sentence, respectively).}
}
\value{
An object of class \code{InputFeatures_FE}.
}
\description{
An InputFeatures object is a single set of (input) features of data used for
(output) feature extraction. Note that this class is similiar to the
InputFeatures class used for simple sequence classification, with annoying
differences. Will eventually standardize; till then, check parameter names.
(RBERT issue #28.)
}
\keyword{internal}
