% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf2-load_checkpoint.R
\name{.load_checkpoint_weights}
\alias{.load_checkpoint_weights}
\title{Load Weights from Checkpoint}
\usage{
.load_checkpoint_weights(mod, ckpt_dir)
}
\arguments{
\item{mod}{BERT model to update with pretrained weights. Modified in
place.}

\item{ckpt_dir}{Directory of checkpoint with pretrained weights.}
}
\value{
TRUE, invisibly. Function is called for side effects.
}
\description{
Updates the given BERT model with the weights from the given checkpoint.
}
\keyword{internal}
