% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_checkpoint.R
\name{.maybe_download_checkpoint}
\alias{.maybe_download_checkpoint}
\title{Find or Possibly Download a Checkpoint}
\usage{
.maybe_download_checkpoint(
  model = c("bert_base_uncased", "bert_base_cased", "bert_large_uncased",
    "bert_large_cased", "bert_large_uncased_wwm", "bert_large_cased_wwm",
    "bert_base_multilingual_cased", "bert_base_chinese", "scibert_scivocab_uncased",
    "scibert_scivocab_cased", "scibert_basevocab_uncased", "scibert_basevocab_cased"),
  dir = NULL,
  ckpt_dir = NULL
)
}
\arguments{
\item{model}{Character vector. Which model checkpoint to download.}

\item{dir}{Character vector. Destination directory for checkpoints. Leave
\code{NULL} to allow RBERT to automatically choose a directory. The path is
determined from the \code{dir} parameter if supplied, followed by the
`RBERT.dir` option (set using \link{set_BERT_dir}), followed by an "RBERT"
folder in the user cache directory (determined using
\code{\link[rappdirs]{user_cache_dir}}). If you provide a \code{dir}, the
`RBERT.dir` option will be updated to that location. Note that the
checkpoint will create a subdirectory inside this \code{dir}.}

\item{ckpt_dir}{The path to the subdir where this checkpoint should
be saved. If model is given, ckpt_dir is inferred.}
}
\value{
TRUE (invisibly)
}
\description{
Verify that the user has a specified checkpoint, and prompt to download if
they don't (in interactive mode).
}
\keyword{internal}
