% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf2-utils.R
\name{.pad_vector}
\alias{.pad_vector}
\title{Pad a Vector to a Certain Length}
\usage{
.pad_vector(x, len, padding)
}
\arguments{
\item{x}{Vector to pad.}

\item{len}{Integer; length to pad or truncate to.}

\item{padding}{Object to use for padding}
}
\value{
`x` padded or truncated to given length.
}
\description{
Pad or truncate the given vector to specified length.
}
\keyword{internal}
