% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_checkpoint.R
\name{.process_BERT_checkpoint}
\alias{.process_BERT_checkpoint}
\title{Unzip and check a BERT checkpoint zip}
\usage{
.process_BERT_checkpoint(dir, checkpoint_zip_path, keep_zip)
}
\arguments{
\item{dir}{Character vector. Destination directory for checkpoint. Leave
`NULL` to allow `RBERT` to automatically choose a directory. The path is
determined from the `dir` parameter if supplied, followed by the
`RBERT.dir` option (set using \link{set_BERT_dir}), followed by an "RBERT"
folder in the user cache directory (determined using
\code{\link[rappdirs]{user_cache_dir()}}). If you provide a `dir`, the
`RBERT.dir` option will be updated to that location.}

\item{checkpoint_zip_path}{The path to which the checkpoint zip should be
downloaded.}

\item{keep_zip}{Logical. Keep the zip file? Leave as \code{FALSE} to save
space.}
}
\value{
`TRUE` invisibly.
}
\description{
Unzip and check a BERT checkpoint zip
}
\keyword{internal}
