% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf2-utils.R
\name{.tf2}
\alias{.tf2}
\title{Check for TensorFlow 2}
\usage{
.tf2()
}
\value{
\code{TRUE} iff TensorFlow installation is version 2.x.
}
\description{
Returns \code{TRUE} iff TensorFlow installation is version 2.x.
}
\keyword{internal}
