% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf2-utils.R
\name{.update_list}
\alias{.update_list}
\title{Update a List}
\usage{
.update_list(list1, list2)
}
\arguments{
\item{list1}{Named list to update.}

\item{list2}{Named list to update with.}
}
\value{
The updated list.
}
\description{
Updates the first list with matching named elements from the second list.
Returns the updated list. Elements in the second list that don't match
(names of) elements in the first list are ignored.
}
\keyword{internal}
