% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_checkpoint.R
\name{download_BERT_checkpoint}
\alias{download_BERT_checkpoint}
\title{Download a BERT checkpoint}
\usage{
download_BERT_checkpoint(model = c("bert_base_uncased",
  "bert_base_cased", "bert_large_uncased", "bert_large_cased",
  "bert_large_uncased_wwm", "bert_large_cased_wwm",
  "bert_base_multilingual_cased", "bert_base_chinese"), dir = NULL,
  url = NULL, force = FALSE, keep_zip = FALSE)
}
\arguments{
\item{model}{Character vector. Which model checkpoint to download.}

\item{dir}{Character vector. Destination directory for checkpoint. Leave
`NULL` to allow `RBERT` to automatically choose a directory. The path is
determined from the `dir` parameter if supplied, followed by the
`RBERT.dir` option (set using \link{set_BERT_dir}), followed by an "RBERT"
folder in the user cache directory (determined using
\code{\link[rappdirs]{user_cache_dir()}}). If you provide a `dir`, the
`RBERT.dir` option will be updated to that location.}

\item{url}{Character vector. An optional url from which to download a
checkpoint. Overrides \code{model} parameter if not NULL.}

\item{force}{Logical. Download even if the checkpoint already exists in the
specified directory? Default \code{FALSE}.}

\item{keep_zip}{Logical. Keep the zip file? Leave as \code{FALSE} to save
space.}
}
\value{
If successful, returns the path to the downloaded checkpoint.
}
\description{
Downloads the specified BERT checkpoint from the Google Research collection.
}
\examples{
\dontrun{
 download_BERT_checkpoint("bert_base_uncased")
 download_BERT_checkpoint("bert_large_uncased")
 temp_dir <- tempdir()
 download_BERT_checkpoint("bert_base_uncased", dir = temp_dir)
}
}
