% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{dropout}
\alias{dropout}
\title{Perform Dropout}
\usage{
dropout(input_tensor, dropout_prob = NULL)
}
\arguments{
\item{input_tensor}{Float Tensor to perform dropout on.}

\item{dropout_prob}{A double giving the probability of dropping out a value
(NOT of KEEPING a dimension as in `tf.nn.dropout`).}
}
\value{
A version of `input_tensor` with dropout applied.
}
\description{
Perform Dropout
}
\examples{
\dontrun{
tfx <- tensorflow::tf$get_variable("none", tensorflow::shape(10L))
dropout(tfx, 0.5)
}
}
