% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_classifier.R
\name{file_based_convert_examples_to_features}
\alias{file_based_convert_examples_to_features}
\title{Convert a set of \code{InputExample}s to a TFRecord file.}
\usage{
file_based_convert_examples_to_features(examples, label_list,
  max_seq_length, tokenizer, output_file)
}
\arguments{
\item{examples}{List of \code{InputExample}s to convert.}

\item{label_list}{Character (or integer?); possible labels for examples.}

\item{max_seq_length}{Integer; the maximum number of tokens that will be
considered together.}

\item{tokenizer}{A tokenizer object to use (e.g. object of class
FullTokenizer).}

\item{output_file}{Character; path to file to write to.}
}
\value{
return value
}
\description{
description
}
\examples{
\dontrun{
}
}
