% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenization.R
\name{FullTokenizer}
\alias{FullTokenizer}
\title{Construct objects of FullTokenizer class.}
\usage{
FullTokenizer(vocab_file, do_lower_case = TRUE)
}
\arguments{
\item{vocab_file}{Path to text file containing list of vocabulary tokens.}

\item{do_lower_case}{Logical: do we convert everything to lowercase?}
}
\value{
An object of class FullTokenizer.
}
\description{
Construct objects of FullTokenizer class.
}
\examples{
\dontrun{
f_tokenizer <- FullTokenizer("vocab.txt", TRUE)
}
}
