% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf2-load_checkpoint.R
\name{get_params_from_checkpoint}
\alias{get_params_from_checkpoint}
\title{Get Model Parameters from Checkpoint}
\usage{
get_params_from_checkpoint(
  ckpt_dir = NULL,
  bert_config_file = find_config(ckpt_dir)
)
}
\arguments{
\item{ckpt_dir}{Character; path to checkpoint directory. If specified, any
other checkpoint files required by this function (\code{vocab_file},
\code{bert_config_file}, or \code{init_checkpoint}) will default to
standard filenames within \code{ckpt_dir}.}

\item{bert_config_file}{Character; the path to a json config file.}
}
\value{
A list of BERT model parameters.
}
\description{
Read the parameters for a BERT model from the specified checkpoint.
}
