% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{get_shape_list}
\alias{get_shape_list}
\title{Return the shape of tensor}
\usage{
get_shape_list(tensor, expected_rank = NULL, name = NULL)
}
\arguments{
\item{tensor}{A tf.Tensor object to find the shape of.}

\item{expected_rank}{The expected rank of \code{tensor}, as an integer vector
or list. If this is specified and the \code{tensor} has a rank not listed
in \code{expected_rank}, an exception will be thrown.}

\item{name:}{Optional name of the tensor for the error message.}
}
\value{
A list of dimensions of the shape of tensor. All static dimensions
  will be returned as native integers, and dynamic dimensions will be
  returned as tf.Tensor scalars. (I'm not very comfortable with this
  behavior. It's not usually good practice to make the return type vary
  depending on the input.)
}
\description{
Returns a list of the shape of tensor, preferring static dimensions. (A
static dimension is known at graph definition time, and a dynamic dimension
is known only at graph execution time.)
https://stackoverflow.com/questions/37096225/
}
\examples{
\dontrun{
with(tensorflow::tf$variable_scope("examples",
                                   reuse = tensorflow::tf$AUTO_REUSE),
 {
   phx <- tensorflow::tf$placeholder(tensorflow::tf$int32, shape = c(4))
   get_shape_list(phx) # static
   tfu <- tensorflow::tf$unique(phx)
   tfy <- tfu$y
   get_shape_list(tfy) # dynamic
 }
)
}
}
