% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_checkpoint.R
\name{\%||\%}
\alias{\%||\%}
\title{Default value for NULL}
\usage{
x \%||\% y
}
\arguments{
\item{x, }{y If x is NULL, will return y; otherwise returns x.}
}
\value{
x or y.
}
\description{
Copied from `rlang` to avoid importing that package.
}
\keyword{internal}
