% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_features.R
\name{input_fn_builder_EF}
\alias{input_fn_builder_EF}
\title{Create an \code{input_fn} closure to be passed to TPUEstimator}
\usage{
input_fn_builder_EF(features, seq_length)
}
\arguments{
\item{features}{A list of features (objects of class
\code{InputFeatures_EF}).}

\item{seq_length}{Integer; the maximum length (number of tokens) of each
example. (Examples should already be padded to this length by this point.)}
}
\value{
An \code{input_fn} closure to be passed to TPUEstimator.
}
\description{
Creates an \code{input_fn} closure to be passed to TPUEstimator. The output
of this closure is the (modified) output of
\code{tensorflow::tf$data$Dataset$from_tensor_slices} (an object of class
"tensorflow.python.data.ops.dataset_ops.BatchDataset"). This function is
similar to \code{input_fn_builder} from run_classifier.R. (RBERT issue #28.)
}
\keyword{internal}
