% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_classifier.R
\name{InputExample}
\alias{InputExample}
\title{Construct objects of class \code{InputExample}}
\usage{
InputExample(guid, text_a, text_b = NULL, label = NULL)
}
\arguments{
\item{guid}{Unique id for the example (character or integer?).}

\item{text_a:}{Character; the untokenized text of the first sequence. For
single sequence tasks, only this sequence must be specified.}

\item{text_b:}{(Optional) Character; the untokenized text of the second
sequence. Only must be specified for sequence pair tasks.}

\item{label:}{(Optional) Character; the label of the example. This should be
specified for train and dev examples, but not for test examples.}
}
\value{
An object of class \code{InputExample}.
}
\description{
An input example is a single training/test example for simple sequence
classification.
}
\examples{
\dontrun{
 input_ex <- InputExample(guid = 0, text_a = "Some text to classify.")
}
}
