% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_classifier.R
\name{InputFeatures}
\alias{InputFeatures}
\title{Construct objects of class \code{InputFeatures}}
\usage{
InputFeatures(
  input_ids,
  input_mask,
  segment_ids,
  label_id,
  is_real_example = TRUE
)
}
\arguments{
\item{input_ids}{Integer Tensor; the sequence of token ids in this example.}

\item{input_mask}{Integer Tensor; sequence of 1s (for "real" tokens) and 0s
(for padding tokens).}

\item{segment_ids}{Integer Tensor; aka token_type_ids. Indicators for which
sentence (or sequence each token belongs to). Classical BERT supports only
0s and 1s (for first and second sentence, respectively).}

\item{label_id}{Integer; represents training example classification labels.}

\item{is_real_example}{Logical; later on this is used as a flag for whether
to "count" this example for calculating accuracy and loss.}
}
\value{
An object of class \code{InputFeatures}.
}
\description{
An InputFeatures object is a single set of features of data.
}
\examples{
\dontrun{
features <- InputFeatures(input_ids, input_mask, segment_ids, label_id)
}
}
