% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenization.R
\name{is_control}
\alias{is_control}
\title{Check whether `char` is a control character.}
\usage{
is_control(char)
}
\arguments{
\item{char}{A character scalar, comprising a single unicode character.}
}
\value{
TRUE if char is a control character.
}
\description{
(R implementation of _is_control from BERT: tokenization.py.)
}
\details{
"\\t", "\\n", and "\\r" are technically control characters but we treat them
as whitespace since they are generally considered as such.
}
\examples{
is_control("\\t")
is_control('c')
}
\keyword{internal}
