% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenization.R
\name{is_punctuation}
\alias{is_punctuation}
\title{Check whether `char` is a punctuation character.}
\usage{
is_punctuation(char)
}
\arguments{
\item{char}{A character scalar, comprising a single unicode character.}
}
\value{
TRUE if char is a punctuation character.
}
\description{
(R implementation of _is_punctuation from BERT: tokenization.py.)
}
\details{
We treat all non-letter/number ASCII as punctuation.
Characters such as "^", "$", and "`" are not in the Unicode
Punctuation class but we treat them as punctuation anyway, for
consistency.
}
\examples{
is_punctuation('.')
is_punctuation('p')
}
\keyword{internal}
