% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenization.R
\name{is_whitespace}
\alias{is_whitespace}
\title{Check whether `char` is a whitespace character.}
\usage{
is_whitespace(char)
}
\arguments{
\item{char}{A character scalar, comprising a single unicode character.}
}
\value{
TRUE if char is a whitespace character.
}
\description{
(R implementation of _is_whitespace from BERT: tokenization.py.)
}
\details{
"\\t", "\\n", and "\\r" are technically control characters but we treat them
as whitespace since they are generally considered as such.
}
\keyword{internal}
