% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{layer_norm_and_dropout}
\alias{layer_norm_and_dropout}
\title{Run layer normalization followed by dropout}
\usage{
layer_norm_and_dropout(input_tensor, dropout_prob = NULL, name = NULL)
}
\arguments{
\item{input_tensor}{Float Tensor to perform layer_norm and dropout on.}

\item{dropout_prob}{A double describing the probability of dropping out a
value (NOT of KEEPING a dimension as in `tf.nn.dropout`).}

\item{name}{Optional variable_scope for layer_norm.}
}
\value{
Tensor resulting from applying layer_norm and dropout to
  \code{input_tensor}.
}
\description{
Run layer normalization followed by dropout
}
\examples{
\dontrun{
tfx <- tensorflow::tf$get_variable("example2", tensorflow::shape(10L))
layer_norm_and_dropout(tfx, dropout_prob = 0.5)
}
}
