% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenization.R
\name{load_vocab}
\alias{load_vocab}
\title{Load a vocabulary file}
\usage{
load_vocab(vocab_file)
}
\arguments{
\item{vocab_file}{path to vocabulary file. File is assumed to be a text
file, with one token per line, with the line number corresponding to the
index of that token in the vocabulary.}
}
\value{
In the BERT Python code, the vocab is returned as an OrderedDict
from the collections package. Here we return the vocab as a named integer
vector. Names are tokens in vocabulary, values are integer indices.
}
\description{
Load a vocabulary file
}
\examples{
\dontrun{
vocab <- load_vocab(vocab_file = "vocab.txt")
}
}
