% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_features.R
\name{make_examples_simple}
\alias{make_examples_simple}
\title{Easily make examples for BERT}
\usage{
make_examples_simple(seq_list)
}
\arguments{
\item{seq_list}{Character vector; text to turn into examples.}
}
\value{
A list of \code{InputExample_EF} objects.
}
\description{
A simple wrapper function to turn a list of text sequences (as a character
vector) into a list of examples suitible for use with RBERT.
}
\examples{
input_ex <- make_examples_simple(c("Here are some words.",
                                   "Here are some more words."))
}
