% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{reshape_from_matrix}
\alias{reshape_from_matrix}
\title{Turn a matrix into a tensor}
\usage{
reshape_from_matrix(output_tensor, orig_shape_list)
}
\arguments{
\item{output_tensor}{Tensor to reshape. What a lousy name for an input.}

\item{orig_shape_list}{Shape to cast Tensor into.}
}
\value{
The Tensor reshaped to rank specified by orig_shape_list.
}
\description{
Reshapes a rank 2 tensor back to its original rank >= 2 tensor. The final
dimension ('width') of the tensor is assumed to be preserved. If a different
width is requested, function will complain.
}
\examples{
\dontrun{
with(
  tensorflow::tf$variable_scope("examples",
    reuse = tensorflow::tf$AUTO_REUSE
  ),
  r2t <- tensorflow::tf$get_variable("r2t",
    dtype = "int32",
    shape = c(10, 20)
  )
)
reshape_from_matrix(r2t, orig_shape_list = c(5L, 2L, 20L))
}
}
