% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{reshape_to_matrix}
\alias{reshape_to_matrix}
\title{Turn a tensor into a matrix}
\usage{
reshape_to_matrix(input_tensor)
}
\arguments{
\item{input_tensor}{Tensor to reshape.}
}
\value{
The Tensor reshaped to rank 2.
}
\description{
Reshapes a >= rank 2 tensor to a rank 2 tensor. The last dimension is
preserved; the rest are flattened.
}
\examples{
\dontrun{
with(
  tensorflow::tf$variable_scope("examples",
    reuse = tensorflow::tf$AUTO_REUSE
  ),
  r3t <- tensorflow::tf$get_variable("r3t",
    dtype = "int32",
    shape = c(10, 20, 3)
  )
)
reshape_to_matrix(r3t)
}
}
