% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_checkpoint.R
\name{set_BERT_dir}
\alias{set_BERT_dir}
\title{Set the directory for BERT checkpoints}
\usage{
set_BERT_dir(dir)
}
\arguments{
\item{dir}{Character vector. Destination directory for checkpoints. Leave
\code{NULL} to allow RBERT to automatically choose a directory. The path is
determined from the \code{dir} parameter if supplied, followed by the
`RBERT.dir` option (set using \link{set_BERT_dir}), followed by an "RBERT"
folder in the user cache directory (determined using
\code{\link[rappdirs]{user_cache_dir}}). If you provide a \code{dir}, the
`RBERT.dir` option will be updated to that location. Note that the
checkpoint will create a subdirectory inside this \code{dir}.}
}
\value{
A list with the previous value of `BERT.dir` (invisibly).
}
\description{
Set a given \code{dir} as the default BERT checkpoint directory for this
session, and create it if it does not exist.
}
\examples{
\dontrun{
set_BERT_dir("fake_dir")
}
}
