% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenization.R
\name{strip_accents}
\alias{strip_accents}
\title{Strip accents from a piece of text.}
\usage{
strip_accents(text)
}
\arguments{
\item{text}{A character scalar, encoded as utf-8.}
}
\value{
text with accents removed.
}
\description{
(R implementation of BasicTokenizer._run_strip_accents from
BERT: tokenization.py.)
}
\examples{
strip_accents("fa\\u00E7ile")
}
\keyword{internal}
