% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenization.R
\name{tokenize.FullTokenizer}
\alias{tokenize.FullTokenizer}
\title{Tokenizer method for objects of FullTokenizer class.}
\usage{
\method{tokenize}{FullTokenizer}(f_tokenizer, text)
}
\arguments{
\item{f_tokenizer}{The FullTokenizer object to apply.}

\item{text}{The text to tokenize.}
}
\value{
A list of wordpiece tokens.
}
\description{
This tokenizer performs some basic cleaning, then splits up text
on whitespace and punctuation.
}
\examples{
\dontrun{
f_tokenizer <- FullTokenizer("vocab.txt", TRUE)
tokenize(f_tokenizer, text = "a bunch of words")
}
}
