% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenization.R
\name{tokenize}
\alias{tokenize}
\title{Generic method for tokenize.}
\usage{
tokenize(x, ...)
}
\arguments{
\item{x}{The Tokenizer object to refer to.}

\item{...}{Additional arguments passed on to methods.}
}
\value{
A list of tokens.
}
\description{
Generic method for tokenize.
}
