% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{transpose_for_scores}
\alias{transpose_for_scores}
\title{Reshape and transpose tensor}
\usage{
transpose_for_scores(
  input_tensor,
  batch_size,
  num_attention_heads,
  seq_length,
  width
)
}
\arguments{
\item{input_tensor}{Tensor to reshape and transpose.}

\item{batch_size}{Size of the first dimension of input_tensor.}

\item{num_attention_heads}{Size of the third dimension of input_tensor. (Will
be transposed to second dimension.)}

\item{seq_length}{Size of the second dimension of input_tensor. (Will be
transposed to third dimension.)}

\item{width}{Size of fourth dimension of input_tensor.}
}
\value{
Tensor of shape: batch_size, num_attention_heads, seq_length, width.
}
\description{
In Python code, this is internal to attention_layer. Pulling it out into
separate function here.
}
\keyword{internal}
