% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_checkpoint.R
\name{.get_model_url}
\alias{.get_model_url}
\title{Get url of a BERT checkpoint}
\usage{
.get_model_url(model = c("bert_base_uncased", "bert_base_cased",
  "bert_large_uncased", "bert_large_cased", "bert_large_uncased_wwm",
  "bert_large_cased_wwm", "bert_base_multilingual_cased",
  "bert_base_chinese"))
}
\arguments{
\item{model}{Character vector. Which model checkpoint to download.}
}
\value{
The url to the specified Google Research BERT model.
}
\description{
Returns the url of the specified BERT checkpoint from the Google Research
collection.
}
\keyword{internal}
