% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.infer_model_paths}
\alias{.infer_model_paths}
\title{Find Paths to Checkpoint Files}
\usage{
.infer_model_paths(
  model = c("bert_base_uncased", "bert_base_cased", "bert_large_uncased",
    "bert_large_cased", "bert_large_uncased_wwm", "bert_large_cased_wwm",
    "bert_base_multilingual_cased", "bert_base_chinese", "scibert_scivocab_uncased",
    "scibert_scivocab_cased", "scibert_basevocab_uncased", "scibert_basevocab_cased"),
  ckpt_dir = NULL,
  vocab_file = find_vocab(ckpt_dir),
  bert_config_file = find_config(ckpt_dir),
  init_checkpoint = find_ckpt(ckpt_dir)
)
}
\value{
A list with components vocab_file, bert_config_file, and
  init_checkpoint.
}
\description{
In some functions, the user can specify a model, a ckpt_dir, and/or specific
paths to checkpoint files. This function sorts all of that out.
}
\details{
This used to inheritParams from extract_features. FIX!
}
\keyword{internal}
