% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_checkpoint.R
\name{.process_BERT_checkpoint}
\alias{.process_BERT_checkpoint}
\title{Unzip and check a BERT checkpoint zip}
\usage{
.process_BERT_checkpoint(dir, checkpoint_archive_path, checkpoint_subdir,
  archive_type)
}
\arguments{
\item{dir}{Character vector. Destination directory for checkpoints. Leave
\code{NULL} to allow RBERT to automatically choose a directory. The path is
determined from the \code{dir} parameter if supplied, followed by the
`RBERT.dir` option (set using \link{set_BERT_dir}), followed by an "RBERT"
folder in the user cache directory (determined using
\code{\link[rappdirs]{user_cache_dir}}). If you provide a \code{dir}, the
`RBERT.dir` option will be updated to that location. Note that the
checkpoint will create a subdirectory inside this \code{dir}.}

\item{archive_type}{How is the checkpoint archived? We currently support
"zip" and "tar-gzip". Leave NULL to infer from the \code{url}.}
}
\value{
\code{TRUE} invisibly.
}
\description{
Unzip and check a BERT checkpoint zip
}
\keyword{internal}
