% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_checkpoint.R
\name{download_BERT_checkpoint}
\alias{download_BERT_checkpoint}
\title{Download a BERT checkpoint}
\usage{
download_BERT_checkpoint(model = c("bert_base_uncased",
  "bert_base_cased", "bert_large_uncased", "bert_large_cased",
  "bert_large_uncased_wwm", "bert_large_cased_wwm",
  "bert_base_multilingual_cased", "bert_base_chinese"),
  destination = "~", url = NULL, overwrite = TRUE)
}
\arguments{
\item{model}{Character vector; which model checkpoint to download.}

\item{destination}{Character vector; destination directory for checkpoint.}

\item{url}{Character vector; a url from which to download a checkpoint.
Overrides \code{model} parameter if not NULL.}

\item{overwrite}{Logical; parameter passed along to \code{utils::unzip}.}
}
\value{
If successful, returns the path to the downloaded checkpoint.
}
\description{
Downloads the specified BERT checkpoint from the Google Research collection.
Will create a "BERT_checkpoints" directory in the specified location (user's
home directory, by default) if not already present.
}
\examples{
\dontrun{
 temp_dir <- tempdir()
 download_BERT_checkpoint("bert_base_uncased", destination = temp_dir)
 download_BERT_checkpoint("bert_large_uncased", destination = temp_dir)
}
}
