% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_classifier.R
\name{file_based_input_fn_builder}
\alias{file_based_input_fn_builder}
\title{summary}
\usage{
file_based_input_fn_builder(x)
}
\arguments{
\item{x}{This parameter will be described when this function is implemented.}
}
\value{
return value
}
\description{
description
}
