% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_files}
\alias{find_files}
\alias{find_vocab}
\alias{find_config}
\alias{find_ckpt}
\title{Find Checkpoint Files}
\usage{
find_vocab(ckpt_dir)

find_config(ckpt_dir)

find_ckpt(ckpt_dir)
}
\arguments{
\item{ckpt_dir}{Character; the path to the checkpoint directory. If this
argument is NULL, the associated functions also return NULL.}
}
\description{
Given the path to a checkpoint directory, return the paths to certain files
in that directory.
}
\section{Functions}{
\itemize{
\item \code{find_vocab}: Find the vocabulary file ('vocab.txt').

\item \code{find_config}: Find the config file ('bert_config.json').

\item \code{find_ckpt}: Find the checkpoint file stub (files begin with
'bert_model.ckpt').
}}

